ops: track "uncommitted" unrefs in `resolutionOp`
When we run through conflict resolution, sometimes we need to unref
blocks that have been pushed to the bserver, but not actually
committed as part of any MD because _another_ conflict was hit before
we could put the MD.  These blocks are special because they still need
to be marked as unreferenced in the next successful PR (so they can be
archived/deleted from the bserver), but without actually counting
their bytes toward the TLF byte usage stored in the MD update.

So this PR introduces a new field in `resolutionOp` to track these
blocks separately, so we can tell whether or not to count their bytes
on a future resolution.

Issue: KBFS-3303