block_tree: fix shifting logic to always update left-side iptrs
If a block got shifted all the way to the left side of an existing
block, and that was its final location, we were failing to update the
new parent pointers to that block.  Instead, update those pointers at
the top of the loop, before breaking when we figure out that it's in
the right location.

Issue: KBFS-3304